wjweb/requirements.txt
configparser==5.0.0
Django==2.1.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
Markdown==3.2.2
mysqlclient==2.0.1
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
whitenoise==5.1.0
